Invert :inherit order for paradox-mode-line-face
With the former :inherit order, if mode-line-buffer-id had a :foreground
property, it would take precedence over font-lock-keyword-face's
:foreground[1].

Reading paradox--build-buffer-id, I get the impression that the idea is
for numbers to be displayed with the regular mode-line-buffer-id face,
while statuses should be highlighted with paradox-mode-line-face.

If mode-line-buffer-id has a :foreground attribute and it has higher
precedence than font-lock-keyword-face, statuses will not look very
different from the rest of the mode line.

Also replace copy-pasted docstring.

[1]: From Info: Emacs Lisp > Display > Faces > Face Attributes:

‘:inherit’
    […] If a list of faces is used, attributes from faces earlier in the
    list override those from later faces.